+++ siddhi-testsuite/src/main/resources/archetype-resources/component/src/test/java/sample/AbstractTemperatureAlertTests.java
@@ -1,4 +1,4 @@
-package sampletests;
+package sample;
 
 import io.siddhi.core.SiddhiAppRuntime;
 import io.siddhi.core.SiddhiManager;
