+++ siddhi-testsuite/src/main/resources/archetype-resources/component/src/test/java/template/__classNameOfAbstractTest__.java
@@ -1,4 +1,4 @@
-package teststemplate;
+package template;
 
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
