+++ siddhi-testsuite/src/main/resources/archetype-resources/component/src/test/java/template/unit/__classNameOfUnitTest__.java
@@ -1,4 +1,4 @@
-package teststemplate.unit;
+package template.unit;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
